Add Elligator2 encoding for ephemeral DH
First attempt at adding the Elligator2 encoding for ephemeral keys, this
adds a lot of code complexity since the elligator2 encoding only needs
to be done for ephemeral keys, and the only implementation available
is ed25519-dallek (or, for now, only a fork of it).

This has been coded only for a Proof of Concept, I'll be happy if anyone
wants to use it, I can help if anyone wants to upstream the changes (but
first the code should be cleaned up), but keep in mind that this commit
is just for experimental purposes.